// src/services/models/Achievement.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Achievement = {
    achievementCondition?: string;
    adminId?: number;
    badgeUrl?: string;
    bannerUrl?: string;
    category?: string;
    conditionType?: string;
    conditionValue?: number;
    createTime?: string;
    description?: string;
    iconUrl?: string;
    id?: number;
    isDelete?: number;
    isHidden?: number;
    isSecret?: number;
    level?: number;
    name?: string;
    points?: number;
    rewardType?: string;
    rewardValue?: string;
    sort?: number;
    updateTime?: string;
};