// src/services/models/AchievementDisplay.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AchievementDisplay = {
    achievementId?: number;
    adminId?: number;
    animationType?: string;
    backgroundColor?: string;
    createTime?: string;
    displayDuration?: number;
    displayType?: string;
    id?: number;
    imageUrl?: string;
    isDelete?: number;
    sort?: number;
    subtitle?: string;
    textColor?: string;
    title?: string;
    updateTime?: string;
};